requirements.txt
python-dotenv
ollama
chromadb
pandas
streamlit
langchain
langchain-community
langchain-ollama